Better homepage sidebar reactivity
Previously the sidebar on the homepage was always 1/5 of the page
no matter the size of the page. This was great, except when the
browser window got really big or became really small. Now if the
browser window gets too small, the sidebar disappear. As the window
gets bigger it scales at 1/5 of the page up until a max of 20rem.