add: css class to be used when rendering markdown
tailwind css adds a bunch of defaults for how `a`, `ul`, `li` and many
more html tags are styled. When we render markdown, we want lists to be
indented and bulleted. We want `a` tags used for link text to render as
a link in coloration and decoration. I'm sure we'll find other markdown
styling desires in the future which will require overriding the tailwind
defaults for many html tags, but this is a good enough start.