update: about page css, switch to vertical viewing on small screens
Previously we were forcing a two column viewing format. This was fine on
larger screens. But on smaller screens everything became incredibly
cramped. Tailwind uses css width minums for delineation. Thus the
default a flex type, except when the width of the viewing window is
greater than tailwind's size `md` (i.e. 768 pixels). When it crosses the
`md` size threshold, it snaps in to the previous two column viewing
style.

Additionally, in the vertical viewing style, the professional
history was for some reason stopping at a width of 506 pixels. I wasn't
sure why this was happening but applying a `w-full` (i.e. `width: 100%`)
to the main div of the professional history div resolved the issue.