Upgrade Phoenix version to 1.7.10
Interestingly I'm getting a bunch of warnings at compile time of the
form
```
warning: no route path for TqmWeb.Router matches ~p"/people/confirm/#{&1}"
  lib/tqm_web/live/person_confirmation_instructions_live.ex:33: TqmWeb.PersonConfirmationInstructionsLive.handle_event/3
```
for a bunch of different route calls. Not sure at the moment what the
problem is. Going to try and solve the issue in the next commit.