Upgrade `phoenix_live_view` version and intertwined deps
Upgrading `phoenix_live_view` required the upgrading of `heroicons` and
`phoenix_live_dashboard`. This is because the versions of those packages
we were one previously dependend on the previous version of
`phoenix_live_view` that we were on.

Additionally, updating these three packages made the warnings we were
seeing go away. So, success!