Move blog_post viewing controls out of blog_post functions
In high security applications I'm generally an advocate of putting the
access calculation as close to the database call as possible. However,
this tends to also make interacting with the server from a shell more
cumbersome. Because this is my personal site, not a commercial site,
nor a FEDRamp or HIPAA application, I think it's reasonable to push
the decision of what content to show to the controller. I may circle
back on this later if actions on thsi website end up having side affect
actions for other features when it makes sense to pass the person
everywhere.