Upgrade to elixir 1.16.0 and erlang 26.2.1 in Dockerfile
Back in c9026c310f7fd093f6a1e37183908a5ac4720023 we upgraded our
development environment to use elixir 1.16.0 and erlang 26.2.1, but
we never upgraded our Dockerfile. This means we've been developing
and deploying on different versions, which isn't great. The good news
is that it appears, everything kept working. That is, I got lucky.
This commit brings our Dockerfile up to date.