Add 20 minute timeout to github workflows
Github action workflows by defualt have a timeout of 360 minutes
(6 hours). This repository's actions should _never_ take that long.
Previously are actions were taking ~2 minutes. Currently, after
upgrading elixir and erlang in https://github.com/QMalcolm/tqm/pull/15
they appear to be taking closer to ~7 minutes. I'm not sure if that is
due to the elixir/erlang upgrade or if github actions are unrelatedly
slower now, as prior to the elixir/erlang upgrade, I hadn't run an
action for this repository in months. Regardless, the repository's
workflows shouldn't take ridiculously long. So I was surprised when I
saw a workflow had been running for _nearly 4 hours_. That shouldn't
happpen. It appears it got hung up on dependencies for some unknown
reason. New jobs are not running into this issue, it appears to be a one
off. None the less, we don't want it to happen again and eat up too many
free "github workflow credits". Thus, a much lower timeout.